+++ src/hooks/useCategoryThemes.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import type { Category } from '@/types/category';
-import { CATEGORY_THEMES_API_URL } from '@/constants/api';
+import { CATEGORY_THEMES_API_URL } from '@/hooks/constants/api';
 
 const fetchCategoryThemes = async (): Promise<Category[]> => {
   const res = await axios.get<{ data: Category[] }>(CATEGORY_THEMES_API_URL);
